+++ changelog.md
@@ -14,6 +14,7 @@ This version is not yet released. If you are reading this on the website, then t
 - Change [`backward ˜`](https://uiua.org/docs/backward)'s glyph back. `𝄈` will continue to work and will be formatted as `˜`.
   - This glyph has much better font support
 - Stabilize [`backward ˜`](https://uiua.org/docs/backward)
+- Add subscripts for [`bits ⋯`](https://uiua.org/docs/bits) to force the number of bits
 - Add sided subscripts for [`reach 𝄐`](https://uiua.org/docs/reach)
 - Add the [`# External!`](https://www.uiua.org/tutorial/documentation#external) semantic comment to mark functions that are provided via Rust code
   - These functions don't require a Uiua implementation and will show up in the LSP